river code is contained in riverScenario removes townScenario rest option removes scenarioBuyFerry removes some unneeded error handling
