+++ checkpointManager.py
@@ -48,57 +48,7 @@ class CheckpointManager:
     #if default choice of -1 is used, use terminal to poll choices
     #if choice isn't -1, then the input comes from a GUI selection and thus a specific scenario can be called directly from the scenario manager
     def riverScenario(self,choice = -1):
-        #if default choice of -1 is used, use terminal to poll choices
-        #if choice isn't -1, then the input comes from a GUI selection and thus a specific scenario can be called directly from the scenario manager
-        if(choice==-1):
-            loopTillValidInput = True
-            while loopTillValidInput:
-                print("You are at the",self.currentCheckpoint.name,"river. What would you like to do?\n1-Attempt to ford the river\n2-Buy a ferry (150$)")
-                try:
-                    playerChoice = int(input())
-                    if(playerChoice == -1):
-                        #debug option
-                        break
-                    if(playerChoice==1):
-                        self.scenarioManager.callScenarioByName("fordRiver")
-                        loopTillValidInput=False
-                    elif(playerChoice==2):
-                        if(self.player.character=="Merchant"):
-                            if(player.money>=self.ferryCost*.9):
-                                loopTillValidInput=False
-                                self.player.money-=self.ferryCost*.9
-                                self.scenarioManager.callScenarioByName("buyFerry")    
-                        else:
-                            if(self.player.money>=self.ferryCost):
-                                self.player.money-=self.ferryCost
-                                loopTillValidInput=False
-                                self.scenarioManager.callScenarioByName("buyFerry")
-                    else:
-                        print("Please try again.")
-                except Exception as e:
-                    print("Invalid Input. Please try again.")
-                    #comment out below as needed
-                    print("Error on line {}".format(sys.exc_info()[-1].tb_lineno))
-                    print(str(e))
-        else:
-            if(choice==1):
-                self.scenarioManager.callScenarioByName("fordRiver")
-            elif(choice==2):
-                if(self.player.character=="Merchant"):
-                    if(self.player.money>=self.ferryCost*.9):
-                        #scenario is called, currently just moves user 10 distance
-                        self.player.money-=self.ferryCost*.9
-                        self.scenarioManager.callScenarioByName("buyFerry")
-                    else: 
-                        print("You don't have enough money for the ferry.")
-                else:
-                    if(self.player.money>=self.ferryCost):
-                        #scenario is called, currently just moves user 10 distance
-                        self.player.money-=self.ferryCost
-                        self.scenarioManager.callScenarioByName("buyFerry")
-                    else: 
-                        print("You don't have enough money for the ferry.")
-
+        self.callScenarioByName("River")
     #if default choice of -1 is used, use a terminal to poll choices
     #if choice isn't -1, then the input comes from a GUI selection and thus a specific scenario can be called directly from the scenario manager
     def townScenario(self,choice = -1):
@@ -116,9 +66,6 @@ class CheckpointManager:
                         self.scenarioManager.callScenarioByName("Travel")
                         loopTillValidInput=False
                     elif(playerChoice==2):
-                        self.scenarioManager.callScenarioByName("rest")
-                        loopTillValidInput=False
-                    elif(playerChoice==3):
                         #
                         self.scenarioManager.callScenarioByName("buySupplies")
                         loopTillValidInput=False
@@ -134,9 +81,6 @@ class CheckpointManager:
                 self.scenarioManager.callScenarioByName("Travel")
                 loopTillValidInput=False
             elif(choice==2):
-                self.scenarioManager.callScenarioByName("rest")
-                loopTillValidInput=False
-            elif(choice==3):
                 #TODO: when the GUI is ready call something to prompt for the amount of food
                 self.scenarioManager.callScenarioByName("buySupplies")
                 loopTillValidInput=False
